agentclient handles multiple waiters for reconnect
When reconnect happens, if there are multiple waiters, all of them
should get notification when the reconnect is done.

AgentClient differentiates between first connect with reconnect. If the
agent fails to connect to agent server, it is considered to be critical,
and we don't retry. After connection is established, we retry connecting
to server to recover from server failure.