Fix conflicting port 8093, problem on local testing.
Both agent and server use port 8093.
Agent has been on that port for 9 months.
Server has been on that port for 3 months.
So opting to fix the server as its the more recent to add the port.
The server only uses the port for admin traffic so should be relatively
safe.
Will flag it in slack.
Fixes issues/140.