Fixed race condition writing to closed channel
Case is when agent gets a packet and attempts to proxy it forward.
However the channel to the connection we are going to forward to
has been closed.
As a result we get an error on writing to the channel.
One possibility is to use a lock to protect the channel for each packet.
Attempting a solution which does not need the lock.
Using a select does not work (https://play.golang.org/p/Az_YCJ24Cu1)
However also shown there is a send made safe using Recover().

Feedback from Chao and Joseph.