Decouple read remote from write to server.
Currently we read from the remote and on the same routine, we write to
the konn server. This means that reading from subsequent reads from
remote are effectively blocked on the write to the konn server. This PR
decouples the read and write. It combines all the writes on to a
channel. Then the writes are on their own routine which reads from that
channel.

Factored in suggested changes from mainred.
Factoring in comment from jkh52 and andrewsykim.
Factoring in comment from mihivagyok.
Renamed writeToKonnServer as writeToProxyServer.