terminate idle connections
It is noticed that the memory consumption of Konnectivity server,
increases when idle connections pile up. The issue is described in more
detail in #311

In order to mitigate the issue, we create a separate goroutine, that
tracks all the connetions per stream and if there is no activity in the
stream after a specified timeout (configured via --grpc-max-idle-time),
we close the stream thereby releasing the resources.

The idea behind this approach is that in a sucessful connection
lifecycle, the apiserver sends a CLOSE_REQ and proxy-server responds
with a CLOSE_RSP but for idle connections this never happens.

In order to to simulate the behaviour of apiserver sending the
CLOSE_REQ, the proxy-server fires up a goroutine and tracks inactive
connections, upon finding so, itself sends a CLOSE_REQ to the same
channel on which it is listening from the api-server.

There are also those connections which are stuck by never receiving the
PacketType_DATA, for those connections, we close the stream once the
timeout occurs.

This also fixes the issue of leaking file descriptors in #276

Signed-off-by: Imran Pochi <imran@kinvolk.io>