Agent: Respect HTTPS_PROXY env vars for proxied connections
Currently, the HTTP{,S}_PROXY env vars are only respected when
connecting to the server but not when creating outgoing connections.
This fixes that.

The usecase for needing this is that controlplane components need to
access services that are only available from within the cluster network
when using a http connect proxy.