Handle broken dial in agent.
When the ANP server gets a request broken in dial, it sends a DIAL_CLS.
This allows the random to be sent back and proper close logic invoked.
CLOSE_REQ is a problem because it uses connection id.
During dial the connection is is zero.
Now extending the ANP agent to also send DIAL_CLS for breakage
during a dial request.
Also added some clean up code for related cases.