agent: smarter readiness check
This commit introduces a smarter way for the readiness check.

Currently as soon as the agent is up, it signals itself as ready
regardless of whether its actually connected to the server or not.

This commit introduces a readiness check in the Agent which reports
true if its connected to atleast one proxy server.

Signed-off-by: Imran Pochi <imranpochi@microsoft.com>