fallback and agent side stop sync
This commit adds a fallback in the case where all the agents in system
are draining. Rather than drop the request with error, we fallback to
the existing behavior i.e continue to the send the request to the agent
even if its draining.

As for the agent side issue, if the agent has sent the DRAIN signal to
the server, ideally it should stop doing the syncOnce with the server.
This mistakes the server the agent is back ready.

Signed-off-by: Imran Pochi <imranpochi@microsoft.com>